Renamed I2CDevice.py to i2c_device.py and updated imports across source
