+++ plugins/lpdpower/lpdpower/quad.py
@@ -1,4 +1,4 @@
-from I2CDevice import I2CException
+from i2c_device import I2CException
 from I2CContainer import I2CContainer
 from tca9548 import TCA9548
 from mcp23008 import MCP23008
